\name{genRal-package}
\alias{genRal-package}
\alias{genRal}
\docType{package}
\title{
  Utilities for data analysis
}
\description{
  This package contains tools and helper functions to perform (usually small)
  manipulations of data structures such as data frames, matrices and lists,
  assist in gene expression data analysis and provide specific functionality
  to analysis scripts.
}
\details{
\tabular{ll}{
  Package: \tab genRal\cr
  Type: \tab Package\cr
  Version: \tab 0.1-0\cr
  Date: \tab 2011-01-07\cr
  License: \tab For internal use only. Please contact the author for more information.\cr
  LazyLoad: \tab yes\cr
}
  The package is meant to be a 'Dependency' for all internal packages
  to provide additional functionality for data analysis in general.
}
\author{
  Vidal Fey (initially)

  Maintainer: Vidal Fey <vidal.fey@utu.fi>
}
\keyword{ package }
