% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.nlines.R
\name{get.nlines}
\alias{get.nlines}
\title{Determine the number of lines in a (large) text file without importing it.}
\usage{
get.nlines(file, n = 1, pattern = NULL, incl.header = FALSE)
}
\arguments{
\item{file}{\code{character of length 1}. File path.}

\item{n}{\code{integer}. Maximum number of lines to read when determining the 'skip' value. Passed to 'get.skip'.}

\item{pattern}{\code{character}. Search pattern to find a certain row in 'file'. Passed to 'get.skip'.}

\item{incl.header}{\code{logical}. Should the file header be included in the count? Length of the header will be determined by 'get.skip'
and the 'pattern' argument. Defaults to 'FALSE'.}
}
\value{
An integer value.
}
\description{
Determine the number of lines in a (large) text file without importing it.
}
