% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.to.list.R
\docType{package}
\name{readmoRe}
\alias{readmoRe}
\title{Utilities for data import}
\description{
A collection of utilities for reading and importing data into R by performing (usually small) manipulations of
    data structures such as data frames, matrices and list and automatically determining import parameters.
}
\details{
\tabular{ll}{
Package: \tab readmoRe\cr
Type: \tab Package\cr
Initial version: \tab 0.1-0\cr
Created: \tab 2011-01-07\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
The main function of the package is \code{read.to.list} which reads a number of different file formats into a list of data objects
    such as data frames, depending on the source file.
}
\author{
Vidal Fey <vidal.fey@gmail.com>
}
\keyword{package}
