\name{rm.empty.cols}
\alias{rm.empty.cols}
\title{
  Remove empty columns from an imported Excel sheet
}
\description{
  \command{rm.empty.cols} removes columns that have only \code{NAs} \emph{AND} whose names
  start with a capital 'X' (unless \option{na.only} is \code{TRUE} in which case all
  \code{NA} columns will be removed).
}
\usage{
  rm.empty.cols(x, na.only = FALSE)
}
\arguments{
  \item{x}{
  a data frame resulting from an imported Excel sheet by means of \command{read.xls}
}
  \item{na.only}{
  logical, should all 'NA' columns be removed and not only those with a column name starting
  with X as generated by Excel (see details section)?
}
}
\details{
  Empty columns in Excel sheets are imported to \code{NA} columns in the resulting
  data frame. Columns that did not have a column name in the spread sheet will result
  in data frame column names starting with 'X'. \command{rm.empty.cols} makes use of these
  two criteria to identify columns that can safely be removed from the data frame.
}
\value{
  A data frame.
}
\author{
  Vidal Fey \email{vidal.fey@utu.fi}
}
\note{
  This is a helper function written in the context of the plate converter utility
  in the \pkg{HTSutil} package. There, it is used in the function \command{read2list}.
}
\seealso{
  \command{\link[HTSutil]{read2list}}
}
\keyword{ utilities }
