\name{se}
\alias{se}
\title{
  Calculate the standard error
}
\description{
  \command{se} calculates the standard error of a numeric vector, matrix or data frame.
}
\usage{
  se(x)
}
\arguments{
  \item{x}{
  numeric vector, matrix or data frame
}
}
\value{
  The standard error.
}
\author{
  Vidal Fey \email{vidal.fey@utu.fi}
}
\seealso{
  \command{\link[stats]{var}}
}
\keyword{ utilities }
