% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sep.R
\name{get.sep}
\alias{get.sep}
\title{Determine field delimiter in text files}
\usage{
get.sep(file, n = 1, pattern)
}
\arguments{
\item{file}{\code{character}. Path name of a text file.}

\item{n}{\code{integer}. Number of lines to be read by \code{readLines}. Defaults to 1.}

\item{pattern}{\code{character}. Search pattern to find a specific line for determining the delimiter.}
}
\value{
If successful, the filed delimiter. If more than on of the possible delimiters is found, an error is returned.
}
\description{
Determine field delimiter in text files
}
\seealso{
\command{\link[base]{readLines}}
}
\keyword{utilities}
