% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.skip.R
\name{get.skip}
\alias{get.skip}
\title{Determine Number of Rows to be Skipped in Text Files}
\usage{
get.skip(file, n = 1, pattern = NULL)
}
\arguments{
\item{file}{(\code{character}). The file name.}

\item{n}{(\code{integer}). The number of lines to be read by \code{readLines()}.}

\item{pattern}{(\code{character}). A search pattern like, e.g., a column  name that is used to find a particular
line in the file to determine the \code{skip} value.}
}
\value{
The \code{skip} value. If no value is determined 0 (zero) is returned.
}
\description{
\command{get.skip} attempts to determine the number of rows that could be skipped when reading text files.
}
\seealso{
\command{\link[base]{readLines}}
}
\keyword{utilities}
