% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.newline.chars.R
\name{rm.newline.chars}
\alias{rm.newline.chars}
\title{Remove 'newline' Characters From Imported Excel Sheets}
\usage{
rm.newline.chars(x, verbose = TRUE)
}
\arguments{
\item{x}{(\code{data.frame}). A data frame resulting from an imported Excel sheet by means of \command{read.xls}.}

\item{verbose}{(\code{logical}). Should verbose output be printed, defaults to \code{TRUE}.}
}
\value{
A data frame.
}
\description{
\command{rm.newline.chars} removes \sQuote{newline} characters (\code{\\n}) from any column of a data frame.
}
\details{
\sQuote{Newline} characters in data frame rows are read verbatim and will cause rows in output
    text files to be distributed across two ore more lines. Such characters, entered accidentally or
    deliberately in the source Excel file, should be avoided. This function removes all \sQuote{newline}
    characters found at the end of a line or replaces them when found within the line text.
}
\keyword{utilities}
